pyproject.toml
[project]
name = "bead"
version = "0.8.1"
description = "Linked frozen computations"
authors = [{name = "Krisztián Fekete"}]
license = {text = "The Unlicense"}
requires-python = ">=3.10"
packages = [
    { include = "bead" },
    { include = "bead_cli" },
    { include = "tracelog.py" },
    { include = "tests" },
]
dependencies = [
    "appdirs",
    "argcomplete>=3.6.2",
    "attrs",
    "cached-property",
]

[project.scripts]
bead = 'bead_cli.main:main'

[dependency-groups]
dev = [
    # tests
    "freezegun",
    "pytest",
    "pytest-cov",

    # debug
    "pysnooper",

    # build
    "pip>=25.1.1",

    # hygiene
    "mypy",
    "pre-commit",
    "ruff",
    "flake8",
    "pytype>=2024.10.11",
]

[tool.flake8]
line_length= 120

[tool.isort]
force_single_line=true
force_sort_within_sections=true
lexicographical=true
order_by_type=true
group_by_package=true

[tool.pyright]
reportUnusedCallResult = false
reportUnknownParameterType = false
reportUnknownVariableType = false
exclude = [
    "executables/**",
    ".pytype/**",
    ".cache/**",
]


[build-system]
requires = ["flit_core"]
build-backend = "flit_core.buildapi"
